add user relations for classrooms and lessons

+++ Services/ClassroomServices/ClassroomService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Data;
 using Dtos.Classroom;
 using Dtos.Lesson;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Models;
 
@@ -15,30 +17,40 @@ namespace Services.ClassroomServices
     {
         private readonly DataContext _context;
         private readonly IMapper _mapper;
-        public ClassroomService(DataContext context, IMapper mapper)
+        private readonly IHttpContextAccessor _contextAccessor;
+        public ClassroomService(DataContext context, IMapper mapper, IHttpContextAccessor contextAccessor)
         {
             _context = context;
             _mapper = mapper;
+            _contextAccessor = contextAccessor;
         }
         public async Task<ServiceResponse<List<GetClassroomDto>>> AddClassroom(AddClassroomDto c)
         {
-            var response = new ServiceResponse<List<GetClassroomDto>>();
             Classroom classroom = _mapper.Map<Classroom>(c);
+            classroom.user = await _context.users.FirstOrDefaultAsync(u => u.id == GetUserId());
             _context.classrooms.Add(classroom);
             await _context.SaveChangesAsync();
-            response.data = await _context.classrooms.Select(c => _mapper.Map<GetClassroomDto>(c)).ToListAsync();
-            return response;
+            return await GetMyClassrooms();
         }
 
         public async Task<ServiceResponse<GetClassroomDto>> AddLessonToClassroom(int classroomId, AddLessonDto l)
         {
             var response = new ServiceResponse<GetClassroomDto>();
-            var dbClassroom = await _context.classrooms.FirstOrDefaultAsync(c => c.id == classroomId);
-            Lesson lesson = _mapper.Map<Lesson>(l);
-            lesson.classroom = dbClassroom;
-            _context.lessons.Add(lesson);
-            await _context.SaveChangesAsync();
-            response.data = _mapper.Map<GetClassroomDto>(dbClassroom);
+            var dbClassroom = await _context.classrooms.FirstOrDefaultAsync(c => c.id == classroomId && c.user.id == GetUserId());
+            if(dbClassroom != null)
+            {
+                 Lesson lesson = _mapper.Map<Lesson>(l);
+                lesson.classroom = dbClassroom;
+                lesson.user = await _context.users.FirstOrDefaultAsync(u => u.id == GetUserId());
+                _context.lessons.Add(lesson);
+                await _context.SaveChangesAsync();
+                response.data = _mapper.Map<GetClassroomDto>(dbClassroom);
+            }
+            else
+            {
+                response.success = false;
+                response.messsage = "Not found!";
+            }
             return response;
 
         }
@@ -69,15 +81,33 @@ namespace Services.ClassroomServices
 
         }
 
+        public async Task<ServiceResponse<List<GetClassroomDto>>> GetMyClassrooms()
+        {
+            var response = new ServiceResponse<List<GetClassroomDto>>();
+            var dbClassrooms = await _context.classrooms.Where(c => c.user.id == GetUserId()).ToListAsync();
+            response.data = dbClassrooms.Select(c => _mapper.Map<GetClassroomDto>(c)).ToList();
+            return response;
+        }
+
         public async Task<ServiceResponse<List<GetClassroomDto>>> RemoveClassroom(int id)
         {
             var response = new ServiceResponse<List<GetClassroomDto>>();
             try
             {
-                Classroom classroom = await _context.classrooms.FirstAsync(c => c.id == id);
-                _context.classrooms.Remove(classroom);
+                Classroom classroom = await _context.classrooms.FirstOrDefaultAsync(c => c.id == id && c.user.id == GetUserId());
+                if(classroom != null)
+                {
+                     _context.classrooms.Remove(classroom);
                 await _context.SaveChangesAsync();
-                response.data = _context.classrooms.Select(c => _mapper.Map<GetClassroomDto>(c)).ToList();
+                response.data = _context.classrooms
+                .Where(c => c.user.id == GetUserId())
+                .Select(c => _mapper.Map<GetClassroomDto>(c)).ToList();
+                }
+                else
+                {
+                    response.success = false;
+                    response.messsage = "Classroom not found!";
+                }
             }
             catch (Exception e)
             {
@@ -92,11 +122,21 @@ namespace Services.ClassroomServices
             var response = new ServiceResponse<GetClassroomDto>();
             try
             {
-                Classroom classroom = await _context.classrooms.FirstOrDefaultAsync(c => c.id == x.id);
-                classroom.title = x.title;
-                classroom.description = x.description;
-                await _context.SaveChangesAsync();
-                response.data = _mapper.Map<GetClassroomDto>(classroom);
+                Classroom classroom = await _context.classrooms
+                .Include(c => c.user)
+                .FirstOrDefaultAsync(c => c.id == x.id && c.user.id == GetUserId());
+                if(classroom != null)
+                {
+                    classroom.title = x.title;
+                    classroom.description = x.description;
+                    await _context.SaveChangesAsync();
+                    response.data = _mapper.Map<GetClassroomDto>(classroom);
+                }
+                else
+                {
+                    response.success = false;
+                    response.messsage = "Classroom not Found";
+                }
             }
             catch (Exception e)
             {
@@ -105,5 +145,7 @@ namespace Services.ClassroomServices
             }
             return response;
         }
+        private int GetUserId() => int.Parse(_contextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+        
     }
 }
